fix json error, fix makefile
- fix a json error, curl output was not well finished
- fix makefile old target detection

+++ maxuxpaste.h
@@ -14,7 +14,7 @@
 
 	typedef struct curl_data_t {
 		char *data;
-		size_t length;
+		unsigned int length;
 		long code;
 		CURLcode curlcode;
